+++ src/assetbundles/registered404s/src/api/registered404s.js
@@ -3,9 +3,6 @@
 import axios from 'axios'
 
 export default {
-    /**
-     * Get cart.
-     */
     get404s(params) {
         axios.defaults.headers.common['X-CSRF-Token'] = Craft.csrfTokenValue
         return axios.post(Craft.getActionUrl('vredirect/catch-all/get-filtered'), params)