+++ src/assetbundles/registered404s/src/app.vue
@@ -33,7 +33,7 @@
             </div>
             <template slot="table-row" slot-scope="props">
                 <span v-if="props.column.field == 'createRedirect'">
-                <button class="btn small" v-on:click="actionCreateRedirect($event, props.row)">Create Redirect</button>
+                <button class="btn small" v-on:click.stop="actionCreateRedirect($event, props.row)">Create Redirect</button>
                 </span>
                 <span v-else>
                 {{props.formattedRow[props.column.field]}}
@@ -48,6 +48,7 @@
     import VueGoodTablePlugin from "vue-good-table"
     import 'vue-good-table/dist/vue-good-table.css'
     import registered404s from "./api/registered404s"
+    import _ from 'lodash'
 
     Vue.use(VueGoodTablePlugin);
 
@@ -80,6 +81,10 @@
                         field: 'uri',
                     },
                     {
+                        label: 'Last Referrer',
+                        field: 'referrer'
+                    },
+                    {
                         label: 'Hits',
                         field: 'hitCount',
                         type: 'number',
@@ -108,10 +113,10 @@
                         filterOptions: {
                             enabled: true,
                             placeholder: 'All',
-                            filterValue: false,
+                            filterValue: "false",
                             filterDropdownItems: [
-                                {value: true, text: 'Only Ignored'},
-                                {value: false, text: 'Only Un-ignored'},
+                                {value: "true", text: 'Only Ignored'},
+                                {value: "false", text: 'Only Un-ignored'},
                             ]
                         },
                     },
@@ -165,12 +170,13 @@
             },
 
             // load items is what brings back the rows from server
-            loadItems() {
+            loadItems: _.debounce(function () {
                 registered404s.get404s(this.serverParams).then(response => {
                     this.totalRecords = response.data.totalRecords;
                     this.rows = response.data.rows;
                 });
-            },
+            }, 500),
+
             actionDelete() {
                 registered404s.delete404s(this.selectedItems).then(() => {
                     this.loadItems();
@@ -191,7 +197,7 @@
             actionCreateRedirect(event, row) {
                 event.preventDefault()
                 event.stopPropagation()
-                window.location = row.createUrl;
+                const child = window.open(row.createUrl);
             },
 
             formatBool(val) {