+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # Redirect Changelog
+### Unreleased
+
+### Added
+- We now keep track of the last referrer for a 404
+
+### Changed
+- Create redirect button on 404 page now opens in a new window
+- You can now "Save and Add Another" redirects
+- You can now press "CMD/CTRL+S" to save the redirect and add another
+
+### Fixed
+- 404 table not respecting "ignored" filter
+- Error saving existing redirect
 
 ## 2.0.11 - 2019-07-03
 ### Added